Handle Redis Sentinel Failover in KombuManager
The redis package raises the exception redis.exceptions.ConnectionError
when a failover occurs as shown by the exception stack trace originating
at message = queue.get(block=True) in _listen(). Since the except block
is only handling OSError and kombu.exceptions.KombuError exceptions,
this error is not detected resulting in loop termination. Add
redis.exceptions.ConnectionError to the except block so the exception is
caught and the new master is detected when connecting to the Sentinel
url as the loop is executed again.

Signed-off-by: Saad Ashfaq <saadashfaq17@gmail.com>